% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asymp_distr_free.R
\name{asymp_distr_free}
\alias{asymp_distr_free}
\title{Computes the asymptotic distribution free test statistic}
\usage{
asymp_distr_free(x, m = 1, weight_matrix_adf = "regular", ...)
}
\arguments{
\item{x}{object of class copstru}

\item{m}{number of expected factors}

\item{weight_matrix_adf}{method to compute weight matrix}

\item{maxit_optim}{the maximum number of iterations in optim}
}
\value{
object of class asymp_distr_free
}
\description{
Computes the asymptotic distribution free test statistic
}

