% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copstruc.R, R/copula_structure_analysis.R
\docType{package}
\name{copstruc}
\alias{copstruc}
\alias{copstruc-package}
\title{copstruc: copula structure analysis}
\usage{
copstruc(x, method = "tail", k = 100, asymp = FALSE, k_select = FALSE,
  ...)
}
\arguments{
\item{x}{data matrix containing n observations of the d-dimensional random variables.}

\item{method}{name of the estimaton method, which can be \code{tail}, \code{kendall}
or \code{standard}.}

\item{k}{number of the largest order statistics, which is used for estimating the
tail dependence function. Default = 1.}

\item{asymp}{if \code{asymp = TRUE}, the asymptotic covariance matrix of lower-diagonal
elements of the copula correlation estimator is calculated. Default = FALSE}

\item{k_select}{logical to decide if copstruc should be used to approximate \code{k}}
}
\value{
object of class copstruc, which is specified by:

  \itemize{
     \item \code{R},	the estimated correlation matrix
     \item \code{Rvec}, the lower diagonal elements of R
     \item \code{Ga}, the asymptotic covariance of Rvec (if \code{asymp == TRUE})
     \item \code{nu}, the estimated coefficient of regular variation (if \code{method == 'tail'})
     \item \code{method}, the estimation method being used
     }
}
\description{
copstruc implements the robust and extreme copulua structure
analysis methods.

\code{copstruc} performs a regular, robust or extreme copula structure
analysis. Currently there is only an assumed factor analysis structure
implemented
}

