% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ellip_tdf.R
\name{ellip_tdf}
\alias{ellip_tdf}
\title{Bivariate elliptical tail dependence function}
\usage{
ellip_tdf(x, y, nu = 2, rho = 0.5)
}
\arguments{
\item{x}{first    argument of the bivariate tail dependence function}

\item{y}{second    argument of the bivariate tail dependence function}

\item{nu}{parameter of regular variation of the generating variable}

\item{rho}{copula correlation}
}
\value{
\code{tde}	value of the tail dependence function
with parameters \code{nu} and \code{rho} at a point (\code{x},\code{y})
}
\description{
ellip_tdf computes the taildependence function corresponding to a
bivariate elliptical copula with parameter \eqn{\nu} and \eqn{\rho}.
It uses the definition with respect to the t-distribution
}

